Incorporated changes for antik
A few words about the repository activity.  There was a
git repository on github.  The antik port was done in an unrelated
repository on the laptop.  To merge repositories, I cloned
the github repository to the laptop, created a laptop branch, and
copied laptop activity to the laptop-antik branch.

For detailed list of changes, see ChangeLog.  But in general, the
major change was to use antik/grid *default...* variables for the
array and float type